fixing code reviewer problems
